+++ src/main/resources/db/migration/V12__create_daily_entrant_count_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE daily_entrant_count (
     daily_entrant_count_id BIGINT AUTO_INCREMENT PRIMARY KEY,
     popup_id BIGINT NOT NULL,
-    entrant_count BIGINT NOT NULL,
+    entrant_count INT NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
 );
